backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI and Machine Learning
transformers==4.35.2
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Hugging Face
datasets==2.15.0
huggingface-hub==0.19.4

# Database
supabase==2.0.3
python-dotenv==1.0.0

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.24.1
requests==2.31.0

# Utilities
pydantic-settings==2.1.0
